requirements.txt
black
Flask
flask-konch
Flask-Migrate
Flask-SQLAlchemy
flask-marshmallow
gunicorn
marshmallow-sqlalchemy
psycopg2-binary
